refactor(mdx): split code into multiple files
